progress on profile pics

+++ projectroot/src/main/java/com/cebedo/pmsys/project/service/ProjectServiceImpl.java
@@ -58,7 +58,7 @@ public class ProjectServiceImpl implements ProjectService {
 
 	@Override
 	@Transactional
-	public Project getByID(int id) {
+	public Project getByID(long id) {
 		return this.projectDAO.getByID(id);
 	}
